criada a session do nextAuth
