+++ .env
@@ -1,0 +1,14 @@
+
+
+DATABASE_URL="mongodb+srv://weitzeltati:Gjm4cq9yP6Hk8AtA@cluster0.uby2r5r.mongodb.net/teste"
+
+
+NEXT_API_URL = http://localhost:8000/api/
+
+
+GOOGLE_ID =559543349856-a5alc9acekigsod335e177ehjb57s31n.apps.googleusercontent.com
+GOOGLE_SECRET =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
+
+
+NEXTAUTH_SECRET='oimlDVS2pMRluzZwE76zZ+zaELZehSPc5oxl+B6QYAI='
+
