+++ src/app/loading.tsx
@@ -1,0 +1,14 @@
+"use client";
+import { useState } from "react";
+import BarLoader from "react-spinners/BarLoader";
+
+const Loading = () => {
+  let [loading, setLoading] = useState(true);
+  return (
+    <div className="flex items-center justify-center h-screen">
+      <BarLoader loading={loading} color="#16a34a" width={100} height={5} />
+    </div>
+  );
+};
+
+export default Loading;
